fix(Combobox): async options highlight (#1854)
* refactor(listbox-root): remove unused parameter from highlightFirstItem function

The function does not use the event in any case.

* fix(combobox-root): highlight first items after item amount is changes

This should fix issues with item highlight while using async loading for options.

* refactor(listbox-virtualizer): fix type-check

* refactor(combobox): use computed instead of reactive with watchers

Using a computed intead of a reactive with updates via watchers
should be more reliable and less prone to issues with stale state.

This also seems to fix the issues we were discussing in this PR.

* fix(combobox-cancel): use filterSearch instead of filterState.search

This happened in a badly resolved merge conflict.

* test(combobox): add story for async options

* fix: virtual list, cleanup

---------

Co-authored-by: Jakub <jakub.michalek@freelo.io>
Co-authored-by: zernonia <zernonia@gmail.com>